fix invitation and approval flow
