+++ app/deprecated-clients/[id]/details-v2/page.tsx
