+++ app/deprecated-clients/[id]/page.tsx
